+++ tests/Unit/MethodsTest.php
@@ -2,6 +2,7 @@
 
 namespace SebastiaanLuca\PipeOperator\Tests\Unit\Classes;
 
+use Closure;
 use PHPUnit\Framework\TestCase;
 
 class MethodsTest extends TestCase
@@ -50,6 +51,32 @@ class MethodsTest extends TestCase
     /**
      * @test
      */
+    public function it can transform a value using a public class method() : void
+    {
+        $this->assertSame(
+            'UPPERCASE',
+            take('uppercase')
+                ->pipe([$this, 'uppercase'])
+                ->get()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function it can transform a value using a private class method() : void
+    {
+        $this->assertSame(
+            'lowercase',
+            take('LOWERCASE')
+                ->pipe(Closure::fromCallable([$this, 'lowercase']))
+                ->get()
+        );
+    }
+
+    /**
+     * @test
+     */
     public function it can transform a value while accepting pipe parameters() : void
     {
         $this->assertSame(
@@ -71,5 +98,25 @@ class MethodsTest extends TestCase
                 ->array_change_key_case(CASE_UPPER)
                 ->get()
         );
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return string
+     */
+    public function uppercase(string $value) : string
+    {
+        return mb_strtoupper($value);
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return string
+     */
+    private function lowercase(string $value) : string
+    {
+        return mb_strtolower($value);
     }
 }